Add spec-driven generator, tests, and CI
