+++ README.md
@@ -1,7 +1,37 @@
 ## Parse Soudan Data Files with Kaitai
 
+### scdms_soudan_spec.py
+Shared source of truth for the Construct schema, the minimal binary payload, and the expected HDF5 values used by both the generator and the tests.
+
+### scdms_soudan_minimal_file.py
+Generates `minimal_sample.soudan` from the spec (small, valid binary for fast testing).
+
 ### scdms_soudan_minimal_file.ipynb
-Create a minimal sample Soudan data file using construct.
+Notebook version of the minimal file generator (kept for exploration).
 
 ### scdms_soudan_parser.ipynb
-Notebook by Adrian to read files using construct and saving them as HDF5
+Notebook by Adrian to read files using construct and saving them as HDF5.
+
+### scdms_soudan_parser.py
+Parses a `.soudan` binary into `minimal_sample.hdf5` using the Construct schema and HDF5 layout.
+
+### test_minimal_hdf5.py
+Validates that `minimal_sample.hdf5` contains the expected datasets, shapes, and values from the spec.
+
+### minimal_sample.soudan / minimal_sample.hdf5
+Tiny sample binary and its parsed HDF5 output, generated by the scripts above.
+
+### Quickstart
+Use the Makefile to regenerate the minimal sample, parse it, and run tests:
+
+```
+make all
+```
+
+Or run individual steps:
+
+```
+make build-soudan
+make build-hdf5
+make test
+```
