+++ requirements.txt
@@ -1,2 +1,4 @@
 construct
 h5py
+numpy
+pytest
